Move auto-sync component to existing color palette, so it can listen to input changes
