+++ plugins/woocommerce/client/admin/client/settings-email/style.scss
@@ -297,7 +297,3 @@ $wc-setting-email-width: 634px;
 	display: inline-flex;
 	gap: 12px;
 }
-
-.wc-settings-email-color-palette-auto-sync {
-	margin: 16px 0;
-}
