+++ plugins/woocommerce/client/admin/client/wp-admin-scripts/settings-embed/index.tsx
@@ -23,7 +23,6 @@ import { registerPaymentsSettingsBannerFill } from '../../payments/payments-sett
 import { registerSiteVisibilitySlotFill } from '../../launch-your-store';
 import { registerBlueprintSlotfill } from '../../blueprint';
 import { registerSettingsEmailColorPaletteFill } from '../../settings-email/settings-email-color-palette-slotfill';
-import { registerSettingsEmailAutoSyncFill } from '../../settings-email/settings-email-auto-sync-slotfill';
 import { registerSettingsEmailImageUrlFill } from '../../settings-email/settings-email-image-url-slotfill';
 import { registerSettingsEmailPreviewFill } from '../../settings-email/settings-email-preview-slotfill';
 import { registerSettingsEmailFeedbackFill } from '~/settings-email/settings-email-feedback-slotfill';
@@ -92,7 +91,6 @@ const registerSlotFills = () => {
 	if ( isFeatureEnabled( 'email_improvements' ) ) {
 		registerSettingsEmailPreviewFill( true );
 		registerSettingsEmailColorPaletteFill();
-		registerSettingsEmailAutoSyncFill();
 		registerSettingsEmailImageUrlFill();
 	} else {
 		registerSettingsEmailPreviewFill( false );
